fix to ticket:1793 - sqlite3 support (only sqlite3 and only in freepbx2.3)
I am using this patch for freepbx 2.2.1, and it does seem to work. if something gets messed up, just email and and I will help you fix this. Tested on 
sqlite3 and MySQL 5 under Debian Etch. I assume this will work for MySQL 4, and if it's not, I will fix it.