Changes for CALL SCREENING in Privacy Manager
Allows a user to select two additional options in the Privacy Manager: Screen Caller - Memory and Screen Caller - No Memory

When memory call screening is selected, the first time a unique caller id calls a user through a DID, the caller must record their 
name which will then be presented to the user this and subsequent calls (via memory in /var/lib/asterisk/sounds/priv-callerintros)

When no memory is selected, each time a caller calls the user, they must record their name.  This is useful if a user gets calls from 
a company with multiple people behind a single caller id.  Choosing this option deletes the priv-callerintros file for that caller id 
prior to adding the "p" option to the Dial command.

Call screening is designed to ignore extension-to-extension calls and ring groups.  Follow-me is a special case and is not considered 
a ring group for this exclusion since call screening would be desirable to a user on follow-me settings.  Locally dialed follow-me 
execution will *not* be screened, though this may be desirable is open for discussion.