Exit with the result code from phpunit
The phpunit.php script is a wrapper on top of the phpunit application.
Previously, we weren't passing the result code from the wrapped application,
so when tests would fail, our wrapper script would pretend everything is ok
and return an exit code of 0. This is particularly bad when you're trying
to unsure unit tests are passing when using an application like Jenkins.
To fix this, we now relay the result code of the phpunit application to
the exit code of the php wrapper script.